Move PendingMessages to its own file
